fix(vite): resolve circular dependency between chunks
