+++ src/CustomProvider/FormattedDate.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useCustom } from './useCustom';
+import { useCustom } from './';
 
 interface FormattedDateProps {
   date: Date;
