+++ src/CustomProvider/index.tsx
@@ -3,8 +3,7 @@ import CustomProvider from './CustomProvider';
 export { FormattedDate } from './FormattedDate';
 export { FormattedNumber } from './FormattedNumber';
 export { useCustom } from './useCustom';
-export { CustomContext } from './CustomProvider';
-export type { CustomProviderProps } from './CustomProvider';
+export { CustomContext, CustomProviderProps } from './CustomContext';
 export type { ReactSuiteComponents } from './types';
 
 export default CustomProvider;
