+++ src/CustomProvider/useCustom.ts
@@ -2,7 +2,7 @@ import { useContext, useCallback, useMemo } from 'react';
 import assign from 'lodash/assign';
 import enGB from '../locales/en_GB';
 import { format, parse, isValid } from '@/internals/utils/date';
-import { CustomContext } from './CustomProvider';
+import { CustomContext } from './CustomContext';
 import type { FormatDateOptions } from '@/internals/utils/date/types';
 import type { ReactSuiteComponents } from './types';
