+++ src/internals/Picker/Listbox.tsx
@@ -18,7 +18,7 @@ import {
   ListChildComponentProps
 } from '@/internals/Windowing';
 import { RSUITE_PICKER_GROUP_KEY } from '@/internals/symbols';
-import { useClassNames, useMount, useEventCallback } from '../hooks';
+import { useClassNames, useMount, useEventCallback } from '@/internals/hooks';
 import { shallowEqual, mergeRefs } from '@/internals/utils';
 import { KEY_GROUP_TITLE } from '@/internals/utils/getDataGroupBy';
 import ListItemGroup from './ListItemGroup';
